gemini/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.62.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
packaging==25.0
proto-plus==1.26.1
protobuf==4.21.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.0.0
requests==2.32.4
rsa==4.9.1
tqdm==4.67.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
nltk==3.8.1

